Update test to add actors and use resusable steps so that the test can loop.
